refactor: 동네 인증 get api 연결 수정
