+++ app/src/main/java/com/issueissyu/fe/data/remote/dto/response/location/LocationBaseResponse.kt
@@ -1,0 +1,5 @@
+package com.issueissyu.fe.data.remote.dto.response.location
+
+data class LocationBaseResponse(
+    val address: String
+)
